+++ lib/core.dart
@@ -5,24 +5,37 @@ import 'dart:convert';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_network_status/network_status.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import 'services.dart';
 
 part 'core/activity.dart';
+
 part 'core/androssy.dart';
+
+part 'core/app.dart';
+
 part 'core/builder.dart';
+
 part 'core/config.dart';
+
 part 'core/controller.dart';
+
 part 'core/errors.dart';
+
 part 'core/fragment.dart';
+
 part 'core/instance.dart';
+
 part 'core/localization.dart';
-part 'core/navigation_bar_config.dart';
+
 part 'core/notifier.dart';
+
 part 'core/provider.dart';
-part 'core/screen_config.dart';
+
 part 'core/settings.dart';
-part 'core/status_bar_config.dart';
+
 part 'core/theme.dart';
+
 part 'core/user.dart';